Move ak init in his own file
And implement it later
