+++ ak/__init__.py
@@ -3,6 +3,7 @@
 from .main import Ak
 from . import ak_sub
 # from . import ak_console
+# from . import ak_init
 from . import ak_db
 from . import ak_other
 from . import ak_build
